refactor(study-session): nest flashcard mode under session-flashcard/
Group the flashcard mode's components into a session-flashcard/ dir with the
orchestrator as index.vue. The modal shell, session-complete, and the shared
deck-context stay at the top level.
