+++ src/components/modals/study-session/session-flashcard/finish-animation.vue
